Borrado de establecimientos

+++ AplicacionesLibres/src/Interfaces/V_Establecimiento.java
@@ -192,7 +192,6 @@ public class V_Establecimiento extends javax.swing.JFrame {
                     recargar();
                     this.dispose();
                 }
-
             }
 
         } else {
@@ -205,9 +204,16 @@ public class V_Establecimiento extends javax.swing.JFrame {
     }//GEN-LAST:event_btn_AceptarActionPerformed
 
     private void btn_borrarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_borrarActionPerformed
-        connEst.insertar("DELETE FROM establecimiento WHERE id_establecimiento='" + txt_ruc.getText() + "'");
-        recargar();
-        this.dispose();
+        System.out.println(String.format("select id_establecimiento from factura where id_establecimiento='%s'",  txt_ruc.getText()) );
+        System.out.println(connEst.verificar_usuario(String.format("select id_establecimiento from factura where id_establecimiento='%s'",  txt_ruc.getText())));
+        if (!connEst.verificar_usuario(String.format("select id_establecimiento from factura where id_establecimiento='%s'",  txt_ruc.getText()))) {
+            connEst.insertar("DELETE FROM establecimiento WHERE id_establecimiento='" + txt_ruc.getText() + "'");
+            recargar();
+            this.dispose();
+        } else {
+            JOptionPane.showMessageDialog(null, "No es posible borrar dicho establecimiento ya que existen facturas ligadas al mismo", "Error", JOptionPane.ERROR_MESSAGE);
+        }
+
     }//GEN-LAST:event_btn_borrarActionPerformed
 
     public void recargar() {
